Add delete all confirmation

+++ main.go
@@ -3,6 +3,7 @@ package main
 import (
 	"errors"
 	"flag"
+	"fmt"
 	"log/slog"
 	"os"
 	"strings"
@@ -23,6 +24,8 @@ type Input struct {
 	DeleteAll bool
 	Delete    bool
 	List      bool
+	// Used to indicate that you are aware of what you're running and don't need confirmation
+	IAmSure bool
 }
 
 func main() {
@@ -33,7 +36,7 @@ func main() {
 	}
 
 	if input.DeleteAll {
-		handleDeleteAll()
+		handleDeleteAll(input)
 	} else if input.List {
 		handleList()
 	} else if input.Delete && input.Bookmark != "" {
@@ -47,7 +50,18 @@ func main() {
 	}
 }
 
-func handleDeleteAll() {
+func handleDeleteAll(input Input) {
+	confirmed := input.IAmSure
+
+	if !confirmed {
+		var input string
+		fmt.Println("Are you sure you want to delete all bookmarks? [y/N]")
+		fmt.Scanln(&input)
+		input = strings.ToLower(input)
+		if input != "y" {
+			return
+		}
+	}
 	err := DeleteBookmarkFile()
 	if err != nil {
 		logger.Error(err.Error())
@@ -131,6 +145,7 @@ func parseFlags() (Input, error) {
 	tags := flag.String("tags", "", "Comma Separated tags for faster searching")
 	search := flag.Bool("search", false, "Search for existing bookmarks against the provided query")
 	deleteAll := flag.Bool("deleteAll", false, "Ignore other flags and delete the bookmark database")
+	iAmSure := flag.Bool("yes", false, "Accept all future confirmations as \"yes\"")
 	d := flag.Bool("delete", false, "Delete the bookmark by the provided key")
 	l := flag.Bool("list", false, "List all bookmarks")
 	t := strings.Split(*tags, ",")
@@ -152,5 +167,6 @@ func parseFlags() (Input, error) {
 		Delete:    *d,
 		Search:    *search,
 		List:      *l,
+		IAmSure:   *iAmSure,
 	}, nil
 }
